Split out token stuff into its own module
as suggested by Gilli. That this was a good decision is reinforced to me
by the fact that one spec that didn't feel like it fit in the lexer
specs suite ended up getting moved to its own suite.

I called this "mmitoken" instead of place "token" because it seems there
is a module in Python's core called "token" and this appeared to be
causing a conflict. Either that, or I misunderstood what was happening.