Added lots of TODOs
because I realized I wasn't going to get this fixed tonight. Take a look
at what I've done and see if it's okay with you. Sorry for jumping in
and just doing a bunch of stuff without asking again... I have so little
time to work on it that when I get a chance, I just go.