Examples of specs as I would like to write them
Using the hypothetical spec/unit-testing module "freeverse" which
supports a hierarchical/nested specs of this form:

    (<description>, <code-clause>, (<sub-specs>))

or for the leaf nodes:

    (<description>, <code-clause>)

or

    should(<description>, <code-clause>)

The top-level tuple is passed into the `spec` function which parses and
registers the spec. The idea is to have largely free-forms specs that
emphasize human readability and structuring the specs in a way that
makes sense for the particular application. They are particularly
designed for input-transform-output type code (i.e. functional) though
they could work in other cases too, conceivably. The code-clauses are
chained, i.e. the output of one is piped into the next. The `should`
function wraps the argument to its code-clause in an object which allows
should-style assertions on it. Also it appends the word "should" to the
beginning of that description of that clause The alternative is to use
the `expect` function for expect-style assertions in a leaf node that
has not been wrapped with should. This should also reduce code
duplication because you can easy form a set of specs around one object
or action and do not have to repeat those since you can break the code
down into "clauses". Also, there is a Result object which is illustrated
below. In case it was not clear above, the piping would be independent
for each leaf node; i.e. if I have a spec of the form,

    spec('5', lambda: 5,
        ('plus', lambda x: lambda y: x + y,
            ('2', lambda plus5: plus5(2), should('be 7', Result.should_be(7))),
            ('5', lambda plus5: plus5(5), should('be 9', Result.should_be(9)))
        )
    )

Then it will execute like this, where '|' represents the result of one
thing being piped in to the next:

    5 | +2 | 7 | passed
    5 | +5 | 10 | failed

And these two execution sequences are independent and could be run in
parallel. The output will look like this:

    5 plus 2 should be 7
    5 plus 5 should be 9

or the like, depending on ouputter. What say you, Gilli?